+++ src/options.ml
@@ -1,4 +1,5 @@
 let default_allow_unused_code_warnings = false
+let default_allow_unused_type_warnings = false
 let perform_checks = false
 
 (* The checks on extensions are only to get better error messages